Add Metrics Group prototype into /prototypes/metrics-groups
