+++ src/js/Routes.jsx
@@ -22,17 +22,17 @@ import Development from 'js/components/development';
 import Callback from 'js/pages/auth/Callback';
 import Waiting from 'js/pages/Waiting';
 import NotFound404 from 'js/pages/NotFound404';
-import InsightLayouts from 'js/pages/insights/InsightLayouts';
 
 import Prototypes from 'js/pages/prototypes';
 import Charts from 'js/pages/prototypes/Charts';
 import DateFilter from 'js/pages/prototypes/DateFilter';
+import MetricGroups from 'js/pages/prototypes/MetricGroups';
 
 export default () => {
 
   const devmode = (
     process.env.NODE_ENV === 'development' ||
-      ['development', 'staging'].includes(window.ENV.environment)
+    ['development', 'staging'].includes(window.ENV.environment)
   );
 
   return (
@@ -79,12 +79,10 @@ export default () => {
             <PrivateRoute path='/waiting' component={Waiting} />
             <Route path='/prototypes/:name?'>
               <Prototypes prototypes={{
-                charts: <Charts />,
-                datefilter: <DateFilter />,
+                'charts': <Charts />,
+                'metrics-groups': <MetricGroups />,
+                'datefilter': <DateFilter />,
               }} />
-            </Route>
-            <Route path='/insights'>
-              <InsightLayouts />
             </Route>
             <Route path='*'>
               <NotFound404 />
